refactor: SelectJdbcTemplate을 JdbcTemplate에 통합
